Refactored the ScriptAdapter to be more basic. Added SchedulerScriptadapter.
